Splitted responses and tasks into separated packages
