+++ src/main/kotlin/com/cognifide/gradle/aem/AemPlugin.kt
@@ -1,6 +1,6 @@
 package com.cognifide.gradle.aem
 
-import com.cognifide.gradle.aem.deploy.*
+import com.cognifide.gradle.aem.deploy.tasks.*
 import com.cognifide.gradle.aem.jar.UpdateManifestTask
 import com.cognifide.gradle.aem.pkg.ComposeTask
 import com.cognifide.gradle.aem.vlt.CheckoutTask
