+++ src/main/kotlin/com/cognifide/gradle/aem/deploy/tasks/AbstractTask.kt
@@ -1,8 +1,14 @@
-package com.cognifide.gradle.aem.deploy
+package com.cognifide.gradle.aem.deploy.tasks
 
 import com.cognifide.gradle.aem.AemConfig
 import com.cognifide.gradle.aem.AemInstance
 import com.cognifide.gradle.aem.AemTask
+import com.cognifide.gradle.aem.deploy.DeployException
+import com.cognifide.gradle.aem.deploy.DeploySynchronizer
+import com.cognifide.gradle.aem.deploy.response.AbstractHtmlResponse
+import com.cognifide.gradle.aem.deploy.response.InstallResponse
+import com.cognifide.gradle.aem.deploy.response.ListResponse
+import com.cognifide.gradle.aem.deploy.response.UploadResponse
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import java.io.File
